test: share code sample format parser
